requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
mutagen>=1.47.0
jinja2>=3.1.4
python-multipart>=0.0.12
pydantic-settings>=2.6.0
musicbrainzngs>=0.7.1
httpx>=0.28.0
itsdangerous>=2.2.0
Pillow>=11.0.0
